+++ iOS/G3MiOSSDK/Commons/Basic/G3MWidget.cpp
@@ -868,7 +868,7 @@ bool G3MWidget::setRenderedSector(const Sector& sector) {
 //  }
 //}
 
-void G3MWidget::changedRendererInfo(const int rendererIdentifier,
+void G3MWidget::changedRendererInfo(const size_t rendererIdentifier,
                                     const std::vector<const Info*>& info) {
   if(_infoDisplay != NULL){
     _infoDisplay->changedInfo(info);
